requirements-dev.txt
tesseract == 0.1.3
requests == 2.24.0
wheel == 0.34.2
pillow == 7.2.0
numpy == 1.18.5
opencv-python == 4.3